+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * **[FIX]** `and X more` message didn't appear properly
 * **[FIX]** Changelog create modal didn't preserve description on close (reported at [Feedbacky](https://app.feedbacky.net/i/changelog-create-modal-issue.4075))
 * **[FIX]** Idea create modal didn't reset chosen tags and added attachment values (reported by [Tigerpanzer_02](https://app.feedbacky.net/i/new-idea-does-not-clear-previosly-used-values.4060))
+* **[FIX]** Inconsistency between changelog update and create, description chars limit was 1800 and 2500.
 
 ### 1.0.0.RC.0
 * **[FEATURE]** Reactions will now display who reacted